move the chilogger from chilogger package to to log package
